update about fixed buggy and error and adding author with qna

+++ src/main/resources/static/js/qna.js
@@ -19,7 +19,6 @@ if (createButton) {
                 qnaContent: document.getElementById("qnaContent").value
             }),
         }).then(() => {
-            alert('문의 글 등록 처리 완료!')
             location.replace("/qna-all")
         }).catch(e => console.error(e));
     })
@@ -39,7 +38,6 @@ if (modifyButton) {
                 qnaContent: document.getElementById("qnaContent").value
             })
         }).then(() => {
-            alert("등록글 수정 처리 완료")
             location.replace(`/qna-all/${id}`)
         })
     })
@@ -47,12 +45,16 @@ if (modifyButton) {
 if (deleteButton) {
     deleteButton.addEventListener('click', () => {
         let id = document.getElementById("qna-id").value
-        fetch(`/api/qna-all/${id}`, {
-            method: 'DELETE',
-        }).then(() => {
-            alert("등록글 삭제 처리 성공")
-            location.replace(`/qna-all`)
-        })
+        if(confirm("진짜로 삭제하시겠어요? 🧐")) {
+            if(confirm("되돌릴 수 없습니다. 🧐")) {
+                fetch(`/api/qna-all/${id}`, {
+                    method: 'DELETE',
+                }).then(() => {
+
+                    location.replace(`/qna-all`)
+                })
+            }
+        }
     })
 }
 document.addEventListener('DOMContentLoaded', function () {
